pyproject.toml
[tool.poetry]
name = "jobapppy"
version = "0.0.1"
description = ""
authors = ["Conor Heine <conor.heine+jobapppy@gmail.com>"]
repository = "https://github.com/cahna/jobapppy"
documentation = "https://cahna.github.io/jobapppy/"
homepage = "https://cahna.github.io/jobapppy/"
readme = "README.md"
keywords = ["resume", "template"]
include = ["jobapppy/templates/*"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Typing :: Typed",
  "Environment :: Console",
  "Topic :: Office/Business",
  "Topic :: Software Development :: Pre-processors",
  "Topic :: Text Processing",
  "Topic :: Utilities",
  "Topic :: Other/Nonlisted Topic", 
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/cahna/jobapppy/issues"

[tool.poetry.scripts]
jobapppy = 'jobapppy.cli:cli'

[tool.poetry.dependencies]
python = "^3.10,<3.12"
pydantic = {version = "^1.10.4", extras = ["email"]}
Jinja2 = "^3.0.1"
typer = {version = "^0.7.0", extras = ["all"]}
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
flake8 = "^6.0.0"
black = "^22.12.0"
isort = "^5.11.4"
pre-commit = "^2.21.0"
tox = "^4.2.4"
mypy = "^0.991"
autoflake = "^2.0.0"
pytest-cov = "^4.0.0"
deptry = "^0.7.1"
coverage = "^7.0.5"
codecov = "^2.1.12"
types-PyYAML = "^6.0.12.12"
types-pygments = "^2.17.0.0"
types-commonmark = "^0.9.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.black]
line-length = 120
target_version = ["py310"]
include = '\.pyi?$'
ignore_missing_imports = ["jinja2"]
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | build
  )/
)
'''

[tool.autoflake]
check = true

[tool.isort]
profile = "black"
src_paths = ["jobapppy", "tests"]
known_first_party = ["jobapppy"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["jobapppy"]
